+++ R/dashboardPage.R
@@ -2,106 +2,109 @@
 #'
 #' Build an adminLTE3 dashboard page
 #'
-#' @param navbar Slot for \link{bs4DashNavbar}.
+#' @param header Slot for \link{bs4DashNavbar}.
 #' @param sidebar Slot for \link{bs4DashSidebar}.
 #' @param body Slot for \link{bs4DashBody}.
 #' @param controlbar Slot for \link{bs4DashControlbar} (right side).
 #' @param footer Slot for \link{bs4DashFooter}.
 #' @param title App title.
-#' @param old_school Whether to use the wonderful sketchy design for Bootstrap 4. FALSE
-#' by default.
-#' @param sidebar_mini Whether to see sidebar icons when \link{bs4DashSidebar}.
-#' is collapsed. TRUE by default.
-#' @param sidebar_collapsed Whether the sidebar is collapsed of not at start. FALSE by default.
-#' @param controlbar_collapsed Whether the sidebar is collapsed of not at start. TRUE by default.
-#' @param controlbar_overlay Whether the sidebar covers the content when expanded.
-#' Default to TRUE.
-#' @param enable_preloader Whether to enable a page loader. FALSE by default.
-#' @param loading_duration Loader duration in seconds. 2s by default.
-#' @param loading_background  Background color during page startup. Blue by default: 
+#' @param freshTheme A skin powered by the fresh package. Not compatible with skin.
+#' See \url{https://dreamrs.github.io/fresh/articles/vars-shinydashboard.html}.
+#' @param preloader bs4Dash uses waiter (see \url{https://waiter.john-coene.com/#/}).
+#' Pass a nested list with 2 elements like \code{list(waiter = list(html = spin_1(), color = "#333e48"), duration = 5)}.
+#' \code{waiter} expects to provide a sub-list to configure \link[waiter]{waiter_show_on_load} (refer to
+#' the package help for all styles). \code{duration} defines the loader timeout.
+#' @param options Extra option to overwrite the vanilla AdminLTE configuration. See 
+#' \url{https://adminlte.io/themes/AdminLTE/documentation/index.html#adminlte-options}.
+#' Expect a list.
 #' \url{https://www.w3schools.com/cssref/css_colors.asp}.
 #'
 #' @examples
 #' if(interactive()){
 #'  library(shiny)
 #'  library(bs4Dash)
+#'  library(fresh)
 #'
-#'  shiny::shinyApp(
-#'    ui = bs4DashPage(
-#'     enable_preloader = TRUE,
-#'     navbar = bs4DashNavbar(),
-#'     sidebar = bs4DashSidebar(),
-#'     controlbar = bs4DashControlbar(),
-#'     footer = bs4DashFooter(),
-#'     title = "test",
-#'     body = bs4DashBody()
+#'  shinyApp(
+#'    ui = dashboardPage(
+#'      freshTheme = create_theme(
+#'       bs4dash_vars(
+#'         navbar_light_color = "#bec5cb",
+#'         navbar_light_active_color = "#FFF",
+#'         navbar_light_hover_color = "#FFF"
+#'       ),
+#'       bs4dash_yiq(
+#'         contrasted_threshold = 10,
+#'         text_dark = "#FFF", 
+#'         text_light = "#272c30"
+#'       ),
+#'       bs4dash_layout(
+#'         main_bg = "#353c42"
+#'       ),
+#'       bs4dash_sidebar_light(
+#'         bg = "#272c30", 
+#'         color = "#bec5cb",
+#'         hover_color = "#FFF",
+#'         submenu_bg = "#272c30", 
+#'         submenu_color = "#FFF", 
+#'         submenu_hover_color = "#FFF"
+#'       ),
+#'       bs4dash_status(
+#'         primary = "#5E81AC", danger = "#BF616A", light = "#272c30"
+#'       ),
+#'       bs4dash_color(
+#'         gray_900 = "#FFF", white = "#272c30"
+#'       )
+#'      ),
+#'      options = NULL,
+#'      header = dashboardHeader(),
+#'      sidebar = dashboardSidebar(),
+#'      body = dashboardBody(
+#'       box(status = "danger"),
+#'       box(status = "primary"),
+#'       box(status = "orange")
+#'      ),
+#'      controlbar = dashboardControlbar(),
+#'      title = "DashboardPage"
 #'    ),
-#'    server = function(input, output) {}
+#'    server = function(input, output) { }
 #'  )
 #' }
 #'
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #'
 #' @export
-bs4DashPage <- function(navbar = NULL, sidebar = NULL, body = NULL, 
-                        controlbar = NULL, footer = NULL, title = NULL,
-                        old_school = FALSE, sidebar_mini = TRUE, 
-                        sidebar_collapsed = FALSE, controlbar_collapsed = TRUE, 
-                        controlbar_overlay = TRUE, enable_preloader = FALSE, 
-                        loading_duration = 2, loading_background = "#1E90FF"){
+bs4DashPage <- function(header, sidebar, body, controlbar = NULL, footer = NULL, title = NULL,
+                        freshTheme = NULL, preloader = NULL, options = NULL){
   
-  sidebar_cl <- NULL
-  if (sidebar_mini) sidebar_cl <- paste0(sidebar_cl, " sidebar-mini")
   
+  # some checks
+  tagAssert(header, type = "nav", class = "main-header")
+  tagAssert(sidebar, type = "aside", class = "main-sidebar")
+  tagAssert(body, type = "div", class = "content-wrapper")
   if (!is.null(controlbar)) {
-    if (!controlbar_overlay) {
-      sidebar_cl <- paste0(sidebar_cl, " control-sidebar-push-slide")
-    } 
+    tagAssert(controlbar[[2]], type = "aside", class = "control-sidebar")
   }
+  if (!is.null(footer)) {
+    tagAssert(footer, type = "footer", class = "main-footer")
+  } 
   
   # create the body content
   bodyContent <- shiny::tags$div(
     class = "wrapper",
-    style = if (enable_preloader) "visibility: hidden; opacity: 0;" else NULL,
-    navbar,
+    header,
     sidebar,
     # page content
     body,
     controlbar,
-    if (!is.null(footer)) footer
+    if (!is.null(footer)) footer,
+    if (!is.null(freshTheme)) {
+      fresh::use_theme(freshTheme)
+    }
   )
   
-  # create the preloader tag if enabled
-  preloaderTag <- if (enable_preloader) {
-    shiny::tags$div(
-      class = "pre",
-      shiny::HTML(
-        '<svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 30 30" enable-background="new 0 0 30 30" xml:space="preserve" width="30" height="30">
-
-		      <rect fill="#FBBA44" width="15" height="15">
-            <animateTransform attributeName="transform" attributeType="XML" type="translate" dur="1.7s" values="0,0;15,0;15,15;0,15;0,0;" repeatCount="indefinite"/>
-		      </rect>	
-
-		      <rect x="15" fill="#E84150" width="15" height="15">
-            <animateTransform attributeName="transform" attributeType="XML" type="translate" dur="1.7s" values="0,0;0,15;-15,15;-15,0;0,0;" repeatCount="indefinite"/>
-		      </rect>	
-      
-		      <rect x="15" y="15" fill="#62B87B" width="15" height="15">
-            <animateTransform attributeName="transform" attributeType="XML" type="translate" dur="1.7s" values="0,0;-15,0;-15,-15;0,-15;0,0;" repeatCount="indefinite"/>
-		      </rect>	
-
-		      <rect y="15" fill="#2F6FB6" width="15" height="15">
-            <animateTransform attributeName="transform" attributeType="XML" type="translate" dur="1.7s" values="0,0;0,-15;15,-15;15,0;0,0;" repeatCount="indefinite"/>
-		      </rect>
-        </svg>
-        '
-      )
-    )
-    
-  }
-  
   # page wrapper
-  shiny::tags$html(
+  shiny::tagList(
     # Head
     shiny::tags$head(
       shiny::tags$meta(charset = "utf-8"),
@@ -115,61 +118,23 @@ bs4DashPage <- function(navbar = NULL, sidebar = NULL, body = NULL,
     ),
     # Body
     add_bs4Dash_deps(
-      theme = old_school,
       shiny::tags$body(
-        class = sidebar_cl,
-        `controlbar-start-open` = tolower(!controlbar_collapsed),
-        `sidebar-start-open` = tolower(!sidebar_collapsed),
-        
-        # set up a time-out for the preloader
-        # The body content disappears first,
-        # then 2s after, the preloader disappears and 
-        # the body content is shown
-        onload = if (enable_preloader) {
-          duration <- loading_duration * 1000
-          paste0(
-            "$(document).ready(function() {
-              setTimeout(function(){
-                $('.pre').remove();
-                $('.wrapper')
-                  .css('visibility', 'visible')
-                  .css('opacity', '');
-                $('body').css('background', '#fff');
-                }, ", duration, ");
-            });
-            "
+        if (!is.null(preloader)) {
+          shiny::tagList(
+            waiter::use_waiter(), # dependencies
+            do.call(waiter::waiter_show_on_load, preloader$waiter)
           )
         },
-        
-        # preloader, if any
-        if (enable_preloader) {
-          shiny::tagList(
-            # preloader CSS
-            shiny::singleton(
-              shiny::tags$style(
-                paste0(
-                  "body {
-                    background: ", loading_background, ";
-                   }
-                  .pre  {
-                    position: absolute;
-                    top: 50%;
-                    left: 50%;
-                    margin-top: -15px;
-                    margin-left: -15px;
-                  }
-                "
-                )
-              )
-            ),
-            # preloader tag
-            preloaderTag,
-            bodyContent
+        onload = if (!is.null(preloader)) {
+          sprintf(
+            "setTimeout(function(){
+            window.loading_screen.finish();
+          }, %s);", preloader$duration * 1000
           )
-        } else {
-          bodyContent
-        }
-      )
+        },
+        bodyContent
+      ), 
+      options = options
     )
   )
 }